requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsons==1.6.3
llama-cloud==0.0.17
llama-index==0.11.9
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.1
llama-index-core==0.11.9
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-ollama==0.3.1
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-faiss==0.2.1
llama-parse==0.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
minijinja==2.2.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
ollama==0.3.3
openai==1.44.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pymongo==4.8.0
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-splitter==1.4
sentence-transformers==3.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
striprtf==0.0.26
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
together==1.2.12
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
typish==1.9.3
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yarl==1.11.1